Expose DecodeFRMPayloadToMACCommands method.
This is needed when a decrypted PHYPayload is marshaled to a slice of
bytes and then unmarshaled back from a slice of bytes to a PHYPayload.
In this case DecryptFRMPayload is not called resulting in the MACCommand
items stored as DataPayload within the FRMPayload field.

One of the cases where this happens is when using the GobEncoder /
GobDecoder (e.g. for storing the PHYPayload into a database) since it
will use the MarshalBinary / UnmarshalBinary methods.